+++ examples/microdata_recode.py
@@ -5,8 +5,8 @@ import piargus as pa
 def main():
     tau = pa.TauArgus(r'C:\Users\LVWN\Desktop\TauArgus4.2.4b2\TauArgus')
     input_df = pd.read_csv('data/microdata.csv')
-    symbol_hierarchy = pa.Hierarchy(["A", "C"])
-    regio_hierarchy = pa.Hierarchy({"Example": ["ExampleDam", "ExampleCity"]})
+    symbol_hierarchy = pa.TreeHierarchy(["A", "C"])
+    regio_hierarchy = pa.TreeHierarchy({"Example": ["ExampleDam", "ExampleCity"]})
     input_data = pa.MicroData(
         input_df,
         hierarchies={'symbol': symbol_hierarchy, "regio": regio_hierarchy},